backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain==0.3.20",
    "openai==1.40.0",
    "pydantic-settings>=2.9.1",
    "qdrant-client==1.13.3",
    "structlog==24.1.0",
    "tiktoken>=0.9.0",
]
